requirements.txt
discord.py>=2.4.0
requests>=2.31.0
playwright>=1.40.0
pandas>=2.0.0
mplfinance>=0.12.10b0
matplotlib>=3.7.0